最大值接口更新

+++ XianHui/Other/Services/NetworkManager.swift
@@ -374,7 +374,7 @@ extension NetworkManager {
     }
     
     //保存报表峰值
-    func saveDailyReportMaxVaule(cashMax:Int,projectmax:Int,prodMax:Int,customerMax:Int,employeemax:Int,completion:DDResultHandler) {
+    func saveDailyReportMaxVaule(cashMax:Float,projectmax:Float,prodMax:Float,customerMax:Float,employeemax:Float,completion:DDResultHandler) {
         let dict:JSONDictionary = [
             "token":Defaults.userToken.value!,
             "cash_amount":cashMax,
